cmd: Move main & use internal
To avoid that the root of the project is the main package, move the main
file inside of the cmd directory and move the sources of the cmd
directory into the new subdirectory cmd/internal.

Having the sources which were in cmd in its own subpackage (no internal)
allows to be able to run main.go without having to add all the list of
Go sources, which will happen if the previous cmd source would have kept
in cmd.